(PC-32289)[API] fix: ensure any OA label is valid
`isVenueAddress` addressed the issue to figure out whether or not
an address is linked to a venue or not.
Therefore, we don't need the "hack" to check whether the label is
equal to the venue's name or not.
Reminder, the venue's OA label is empty in the database BUT
is filled with venue's common name which we get back when the
front sends its data.