(BSR)[API] chore: Avoid n+1 query
venueId was accessed through the collectiveDmsApplication
instance. The thing is, collectiveDmsApplication are originally
fetched joining over their Venue. Which means:

- Venue.id -> no n+1 query
- Venue.collectiveDmsApplication[0] -> no n+1 query
- Venue.collectiveDmsApplication[0].venue.id -> n+1 query

We were issuing this n+1 query for every collectiveDmsApplication
of every Venue.

Since the Venue is the table used in the FROM clause, we cannot
reasonably joinedload the venue along side the
collectiveDmsApplication, that would mean fetching the Venue.id
column twice.
However, since the Venue is primarly fetched, we can simply pass the
Venue.id to the DMSApplicationForEAC serializer.

As DMSApplicationForEAC is never used on its own but always as part of an
outer serializer (which also systematically contains the Venue.id)
i think this field should be removed and the venueId retrieve from
the outer serializer. But if it's really, really needed, we now have it without
any overhead.