(PC-34142)[API] fix: edit stock: compare dates without tz
When editing a stock, the existing and the new beginning times were
compared but could never be equal since the first does not have any TZ
information (datetime is saved without it inside the DB) and the later
does.

Now, to find out if the beginning date needs to be updated, the check is
performed with two TZ-unaware datetime objects.

Of course, this should not be needed: the real fix is to save the TZ
inside the database.

So please, please, please
Let me, let me, let me
Let me get the timezones that I want
This time