(PC-37108)[API] feat: pro: venue shares offers' openingHours models
A recent commit added opening hours to offers, only using the PRO app.
It has one main drawback: the database table used is the same, the main
logic is (should be) the same, but the pydantic models are different
(because it is (way) less permissive).

This commit fixes that: both venues and offers should use the same
pydantic models when it comes to openin hours.

Notes:
  * the commit edits some small portion of frontend code
    (dont-known-what-Im-doing-but-it-seems-to-work style);
  * the commit moved core.offers.opening_hours modules to
  core.opening_hours (which makes more sens);
  * the commit modified some opening_hours.api functions to
  allow change tracking that is needed by `edit_venue`.