(PC-37678)[API] feat: pro: add offer.canBeEvent field
Meaning: the offer's subcategory is an event one.

Why? The Offer.isEvent definition will soon be updated from: an offer
whose subcategory is an event to and offer with at least one stock that
has a beginningDatetime set (and an event subcategory). Without this new
field, the offer creation steps would break because the front end would
assume that the offer could not be an event until the stocks step is
reached (which depends on wether the offer could be an event or not)...
which would make things hard since the step is not the same for things
and events...