(PC-37825)[API] feat: rename isDigital to hasUrl
Goal: isDigit will be redefined. To avoid any confusion add setup a
smooth transition, this new field will keep the old behaviour and the
new one will be added only when needed (after).

Context: before a digital offer is an offer with an url, now it is an
offer whose subcategory uses an online platform.