(PC-38919)[API] fix: defer ubble id archive after the transaction
Using `@atomic` decorator prevents the transaction from committing until
the beneficiary activation is done, which can take a bit of time. During
that time, the cloud task can be called too soon and crash because of
unexpected state.

The unit tests didn't catch that bug because the cloud tasks are called
eagerly and share the only SQLAlchemy session: the modifications do not
need to be committed for the task to see the new state.